% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Myfunction.R
\name{bw_density_test}
\alias{bw_density_test}
\title{Density Test}
\usage{
bw_density_test(X, Y, alpha, bw)
}
\arguments{
\item{X}{the first sample(dim=1)}

\item{Y}{the second sample(dim=1)}

\item{alpha}{the confidence coefficient}

\item{bw}{the bandwidth,only use in bw_density_test}
}
\value{
The list of result(1 stand for equal,0 stand for inequal),Tstatistic,Standard
}
\description{
Test two samples have the same density or not
}
\examples{
\dontrun{
 n1 <- 30
 n2 <- 40
 X1 <- rnorm(n1,0,1)
 X2 <- rnorm(n2,0,1)
 bw <- 0.49
 bw_density_test(X1,X2,0.05,0.49)
}
}
