% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Myfunction.R
\name{bw_Independence_test}
\alias{bw_Independence_test}
\title{Independence Test}
\usage{
bw_Independence_test(X, Y, alpha, bw1, bw2)
}
\arguments{
\item{X}{the first sample(dim=1)}

\item{Y}{the second sample(the same sample size with X,dim=1)}

\item{alpha}{the confidence coefficient}

\item{bw1}{the bandwidth of X,only use in bw_Independence_test}

\item{bw2}{the bandwidth of Y,only use in bw_Independence_test}
}
\value{
The list of result(independent,dependent,the different sample size),T-statistic,Standard
}
\description{
Test two samples(the same sample size) are indenpendent or not
}
\examples{
\dontrun{
 n1 <- 30
 X1 <- rnorm(n1,0,1)
 X4 <- X1^2
 h1 <- 0.687
 h4 <- 1.262
 bw_Independence_test(X1,X4,0.05,h1,h4)
}
}
