% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CLap}
\alias{CLap}
\title{Laplace sample by Rcpp}
\usage{
CLap(sigma, x0, N)
}
\arguments{
\item{sigma}{the SD of Normal distribution}

\item{x0}{the first data}

\item{N}{the sample size}
}
\value{
a list,x stands for random sample of Laplace sample,k is the acceptance number
}
\description{
Generate the Laplace sample by Rcpp,using the normal distribution
}
\examples{
\dontrun{
res <- CLap(sigma,x0,N)
}
}
