% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Myfunction.R
\name{density_test}
\alias{density_test}
\title{Density Test}
\usage{
density_test(X, Y, alpha)
}
\arguments{
\item{X}{the first sample(dim=1)}

\item{Y}{the second sample(dim=1)}

\item{alpha}{the confidence coefficient}
}
\value{
The list of result(1 stand for equal,0 stand for inequal),Tstatistic,Standard
}
\description{
Test two samples have the same density or not
}
\examples{
\dontrun{
 n1 <- 30
 n2 <- 40
 X1 <- rnorm(n1,0,1)
 X2 <- rnorm(n2,0,1)
 density_test(X1,X2,0.05)
}
}
