% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Myfunction.R
\name{f1}
\alias{f1}
\title{f1}
\usage{
f1(X, h, m)
}
\arguments{
\item{X}{the first sample(dim=1)}

\item{h}{the bandwidth}

\item{m}{the number of the leave-out}
}
\value{
The function value
}
\description{
helper function
}
