% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Myfunction.R
\name{f2}
\alias{f2}
\title{f2}
\usage{
f2(X, Y, h1, h2, m)
}
\arguments{
\item{X}{the first sample(dim=1)}

\item{Y}{the second sample(dim=1)}

\item{h1}{the bandwidth of X}

\item{h2}{the bandwidth of Y}

\item{m}{the number of the leave-out}
}
\value{
The function value
}
\description{
helper function
}
