% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Myfunction.R
\name{fhat_adaptive}
\alias{fhat_adaptive}
\title{Adaptive estimate density}
\usage{
fhat_adaptive(Sample, x, h)
}
\arguments{
\item{Sample}{the sample}

\item{x}{the function's argument}

\item{h}{the initial bandwidth}
}
\value{
the estimate density function
}
\description{
Compute the adaptive estimate density by computing the suitable bandwidth
}
\examples{
\dontrun{
set.seed(2020)
rsample <- function(N){
 lnY <- rnorm(N,mean = 0,sd=1)
 Y <- exp(lnY)
 return(Y)
}
true_density <- function(t){
 return_value <- dnorm(log(t),0,1)/t
 return(return_value)
}
h <- 0.3
n <- 200
X <- rsample(n)
y <- seq(0.002,16,0.002)
plot(y, true_density(y), "l", main = "adaptive")
lines(y,fhat_adaptive(X,y,h),"l",col=2)
}
}
