% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Myfunction.R
\name{Independence_test}
\alias{Independence_test}
\title{Independence Test}
\usage{
Independence_test(X, Y, alpha)
}
\arguments{
\item{X}{the first sample(dim=1)}

\item{Y}{the second sample(the same sample size with X,dim=1)}

\item{alpha}{the confidence coefficient}
}
\value{
The list of result(independent,dependent,the different sample size),T-statistic,Standard
}
\description{
Test two samples(the same sample size) are indenpendent or not
}
\examples{
\dontrun{
 n1 <- 30
 n2 <- 40
 X1 <- rnorm(n1,0,1)
 X2 <- rnorm(n2,0,1)
 Independence_test(X1,X2,0.05)
}
}
