% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Myfunction.R
\name{n_digit}
\alias{n_digit}
\title{n_digit}
\usage{
n_digit(a, n)
}
\arguments{
\item{a}{the Postive Integer}

\item{n}{the number of digit}
}
\value{
the function value
}
\description{
From the blog(https://www.cnblogs.com/onepixel/articles/7674659.html)
}
