% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Myfunction.R
\name{Sort_bubble}
\alias{Sort_bubble}
\title{Sort_bubble}
\usage{
Sort_bubble(X)
}
\arguments{
\item{X}{the array needed to sort}
}
\value{
the array has been sorted
}
\description{
From the blog(https://www.cnblogs.com/onepixel/articles/7674659.html)
}
\examples{
\dontrun{
X <- sample(1:50,size=20,replace=T)
X
Sort_bubble(X)
}
}
