\name{StatComp20077-package}
\alias{StatComp20077-package}
\docType{package}
\title{
  sorting algorithm,computing estimate function,test of 2 samples
}
\description{
1 nine types of sorting algorithm
2 using the kernel method to test the density of two samples are same or not
3 using the kernel method to test two samples are independent or not
4 The function to compute the adaptive estimate function
}
\details{
  1 Sort_counting_Integer only use for integer
  2 Sort_radix_PostiveInteger only use for postive integer
  3  When test two samples,because of the bandwidth is difficult to choose,the result is not very good.Users can choose the bandwidth by himself in a similar function(Only add an addtional para bw,others are same)
}
\author{
Peiran Liu

Maintainer: Peiran Liu 
<lpr1423@mail.ustc.edu.cn>
}

\keyword{ package }
