% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Myfunction.R
\name{Ts}
\alias{Ts}
\title{Tstat}
\usage{
Ts(X, Y, h)
}
\arguments{
\item{X}{the first sample(dim=1)}

\item{Y}{the second sample(dim=1)}

\item{h}{the bandwidth}
}
\value{
The test statistic
}
\description{
helper function
}
