% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Myfunction.R
\name{Ts2}
\alias{Ts2}
\title{T-stat}
\usage{
Ts2(X, Y, h1, h2)
}
\arguments{
\item{X}{the first sample(dim=1)}

\item{Y}{the second sample(dim=1)}

\item{h1}{the bandwidth of X}

\item{h2}{the bandwidth of Y}
}
\value{
T-statistic
}
\description{
helper function
}
